@@ -2,8 +2,10 @@ package com.whois.whoiswho.activity;
 
 import android.app.Activity;
 import android.app.Fragment;
+import android.app.FragmentManager;
 import android.content.ActivityNotFoundException;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.design.widget.NavigationView;
@@ -24,7 +26,10 @@ import com.squareup.picasso.Picasso;
 import com.whois.whoiswho.R;
 import com.whois.whoiswho.model.User;
 import com.whois.whoiswho.screens.edit_profile.EditProfileFragment;
+import com.whois.whoiswho.screens.login.LoginFragment;
+import com.whois.whoiswho.screens.mainscreen.TabFragment;
 import com.whois.whoiswho.utils.KeyboardManager;
+import com.whois.whoiswho.utils.SharedPreferencesSaver;
 import com.whois.whoiswho.utils.Toaster;
 
 import java.util.ArrayList;
@@ -61,16 +66,12 @@ public class MainActivity extends Activity implements MainActivityMethods, Navig
         phoneTextView = (TextView)headerView.findViewById(R.id.phoneTextView);
         avatarImageView = (CircleImageView)headerView.findViewById(R.id.avatarImageView);
         editProfileButton = (ImageView)headerView.findViewById(R.id.editProfileButton);
-        editProfileButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-               // getFragmentManager().beginTransaction().setCustomAnimations(0, 0, 0, R.anim.slide_out_to_top).add(R.id.flContent, new EditProfileFragment()).addToBackStack(null).commit();
-                getFragmentManager().beginTransaction().setCustomAnimations(0,0,0, R.animator.slide_out_to_bottom).add(R.id.flContent, new EditProfileFragment()).addToBackStack(null).commit();
-                drawer.closeDrawer(GravityCompat.START);
-            }
+        editProfileButton.setOnClickListener(v -> {
+            getFragmentManager().beginTransaction().setCustomAnimations(0,0,0, R.animator.slide_out_to_bottom).add(R.id.flContent, new EditProfileFragment()).addToBackStack(null).commit();
+            drawer.closeDrawer(GravityCompat.START);
         });
         navigationView.setNavigationItemSelectedListener(this);
-}
+    }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
@@ -94,13 +95,23 @@ public class MainActivity extends Activity implements MainActivityMethods, Navig
         return super.onOptionsItemSelected(item);
     }
 
+    public void initActivity() {
+        getFragmentManager().popBackStack(null, FragmentManager.POP_BACK_STACK_INCLUSIVE);
+        if (!TextUtils.isEmpty(SharedPreferencesSaver.get().getToken())) {
+            switchFragment(new TabFragment(), false);
+        } else {
+            drawer.setDrawerLockMode(DrawerLayout.LOCK_MODE_LOCKED_CLOSED);
+            switchFragment(new LoginFragment());
+            //view.switchFragment(new TabFragment(), false);
+        }
+    }
+
     public void displayUserInfo(User user) {
         nameTextView.setText(user.getFirstname() + " " + user.getSurname());
         phoneTextView.setText(PhoneNumberUtils.formatNumber(user.getNumber(), user.getCountyIso()));
         if (!TextUtils.isEmpty(user.getAvatarPath())) {
             Picasso.with(getApplicationContext()).load(user.getAvatarPath()).into(avatarImageView);
         }
-
     }
 
 
@@ -111,11 +122,10 @@ public class MainActivity extends Activity implements MainActivityMethods, Navig
 
     @Override
     public void switchFragment(Fragment fragment, boolean addToFragmentList) {
-        if (addToFragmentList) {
-            getFragmentManager().beginTransaction().add(R.id.flContent, fragment).addToBackStack(null).commit();
-        } else {
-            getFragmentManager().beginTransaction().add(R.id.flContent, fragment).commit();
-        }
+        if (addToFragmentList)
+            getFragmentManager().beginTransaction().replace(R.id.flContent, fragment).addToBackStack(null).commit();
+        else
+            getFragmentManager().beginTransaction().replace(R.id.flContent, fragment).commit();
     }
 
     @Override
@@ -159,7 +169,9 @@ public class MainActivity extends Activity implements MainActivityMethods, Navig
 
     @Override
     public void onRequestPermissionsResult(int requestCode, String permissions[], int[] grantResults) {
-
+        if (grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+            initActivity();
+        }
     }
 
     @Override
