@@ -29,19 +29,9 @@ public class MainActivityPresenter {
                 (ContextCompat.checkSelfPermission(view, Manifest.permission.PROCESS_OUTGOING_CALLS) != PackageManager.PERMISSION_GRANTED)&&
                 (ContextCompat.checkSelfPermission(view, Manifest.permission.READ_CALL_LOG) != PackageManager.PERMISSION_GRANTED)&&
                 (ContextCompat.checkSelfPermission(view, Manifest.permission.WRITE_CONTACTS) != PackageManager.PERMISSION_GRANTED)) {
-            view.switchFragment(new LoginFragment());
-            view.drawer.setDrawerLockMode(DrawerLayout.LOCK_MODE_LOCKED_CLOSED);
             ActivityCompat.requestPermissions(view, new String[]{Manifest.permission.PROCESS_OUTGOING_CALLS, Manifest.permission.READ_CONTACTS, Manifest.permission.WRITE_CONTACTS, Manifest.permission.READ_CALL_LOG, Manifest.permission.CALL_PHONE, Manifest.permission.SYSTEM_ALERT_WINDOW}, view.MY_PERMISSIONS_REQUEST);
         }
-        else {
-            if (!TextUtils.isEmpty(SharedPreferencesSaver.get().getToken())) {
-                view.switchFragment(new TabFragment(), false);
-            } else {
-                view.drawer.setDrawerLockMode(DrawerLayout.LOCK_MODE_LOCKED_CLOSED);
-                view.switchFragment(new LoginFragment());
-                //view.switchFragment(new TabFragment(), false);
-            }
-        }
+        view.initActivity();
     }
 
 
